backend/requirements.txt
fastapi
uvicorn
celery
redis
openai
python-dotenv
pandas
python-docx
owlready2
